requirements_v4.txt
google-genai
PyPDF2
pdf2image
Pillow
python-dotenv 
poppler